--- src/scoreboard.py
@@ -1,28 +1,0 @@
-from src.setting import *
-
-
-class Scoreboard():
-    def __init__(self,x=-1,y=-1):
-        self.score = 0
-        self.tempimages,self.temprect = load_sprite_sheet('numbers.png',12,1,15,int(15*6/5),-1)
-        self.image = pygame.Surface((80,int(15*6/5)))
-        self.rect = self.image.get_rect()
-        if x == -1:
-            self.rect.left = width*0.89
-        else:
-            self.rect.left = x
-        if y == -1:
-            self.rect.top = height*0.05
-        else:
-            self.rect.top = y
-
-    def draw(self):
-        screen.blit(self.image,self.rect)
-
-    def update(self,score):
-        score_digits = extractDigits(score)
-        self.image.fill(background_col)
-        for s in score_digits:
-            self.image.blit(self.tempimages[s],self.temprect)
-            self.temprect.left += self.temprect.width
-        self.temprect.left = 0
